[Refactor] Split functions out to file
